Removal of some compilation warnings.

+++ src/shared_processing/core/asResultsAnalogsDates.cpp
@@ -1,17 +1,17 @@
-/** 
+/**
  *
  *  This file is part of the AtmoSwing software.
  *
- *  Copyright (c) 2008-2012  University of Lausanne, Pascal Horton (pascal.horton@unil.ch). 
+ *  Copyright (c) 2008-2012  University of Lausanne, Pascal Horton (pascal.horton@unil.ch).
  *  All rights reserved.
  *
- *  THIS CODE, SOFTWARE AND INFORMATION ARE PROVIDED "AS IS" WITHOUT WARRANTY  
+ *  THIS CODE, SOFTWARE AND INFORMATION ARE PROVIDED "AS IS" WITHOUT WARRANTY
  *  OF ANY KIND, EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
  *  IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR
  *  PURPOSE.
  *
  */
- 
+
 #include "asResultsAnalogsDates.h"
 
 #include "asFileNetcdf.h"
@@ -381,7 +381,7 @@ bool asResultsAnalogsDates::Load(const wxString &AlternateFilePath)
 
     // Get data
     size_t IndexStart[2] = {0,0};
-    size_t IndexCount[2] = {Ntime, Nanalogs};
+    size_t IndexCount[2] = {size_t(Ntime), size_t(Nanalogs)};
     ncFile.GetVarArray("analogsdates", IndexStart, IndexCount, &analogsDates[0]);
     ncFile.GetVarArray("analogscriteria", IndexStart, IndexCount, &analogsCriteria[0]);
